Trying to setup seamless Share clustering.
It does not work. I have two shares with sticky sessions, but
if the active Share instance goes down, and the page is reloaded
we have to login again. This seems to indicate Session sync is
not working or that Nginx is behaving poorly when it has to
failover. Could work with curl/wget and a cookie jar to try and
validate this further.